+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/theming/AppTheme.kt
@@ -1,0 +1,19 @@
+package com.altmann.choresmanager.theming
+
+import androidx.compose.ui.graphics.Color
+
+data class AppTheme(
+    val primary: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val onPrimary: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val secondary: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val onSecondary: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val background: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val onBackground: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val primaryContainer: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val onPrimaryContainer: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val surface: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val onSurface: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val surfaceVariant: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val onSurfaceVariant: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+    val error: Color = Color.hsl(0.5f, 0.5f, 0.5f),
+)